# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DallyLetter Elidems - CAPS Education Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/block:
    patch:
      operationId: blockUser
      tags: [users]
      summary: Block or unblock a user (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/promote:
    patch:
      operationId: promoteUser
      tags: [users]
      summary: Promote/demote user role or assign prefect (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # LESSONS ROUTES
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, teacher, owner]
        isPrefect:
          type: boolean
        isBlocked:
          type: boolean
        phone:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        lastPaymentDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - isPrefect
        - isBlocked
        - createdAt
...
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    BlockUserBody:
      type: object
      properties:
        isBlocked:
          type: boolean
      required:
        - isBlocked
    PromoteUserBody:
      type: object
      properties:
        role:
          type: ["string", "null"]
          enum: [student, teacher, owner, null]
        isPrefect:
          type: ["boolean", "null"]
...